backend/requirements.txt
# Django packages
Django==5.0.4
djangorestframework==3.15.0
django-cors-headers==4.3.1
django-filter==24.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
dj-rest-auth==5.0.2

# Scraping
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1
webdriver-manager==4.0.1
lxml==5.1.0

# Async processing
celery==5.3.6
redis==5.0.1
django-celery-results==2.5.1
django-celery-beat>=2.5.0


# Development tools
pytest==8.1.1
pytest-django==4.8.0
coverage==7.4.3
black==24.2.0
isort==5.13.2
flake8==7.0.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Other utils
Pillow==10.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
loguru==0.7.2
colorlog==6.9.0
concurrent-log-handler==0.9.25
